requirements.txt
pandas==1.5.3
numpy==1.26.4
scikit-learn==1.7.2
pyarrow==14.0.2

rectools==0.17.0
implicit==0.7.2

fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
requests==2.32.5

clickhouse-connect==0.10.0
psycopg2-binary==2.9.9
sqlalchemy==1.4.54

kafka-python==2.3.0
streamlit==1.52.2

prometheus-fastapi-instrumentator==7.1.0
loguru==0.7.3
python-dotenv==1.0.0
